fix: on_connect resets, alarm timeout tracking, dead code cleanup
Code review round 2 fixes:
- Fix on_connect MOSFET/CW reset to use inverted value (not false) so
  entities re-publish after MQTT reconnect even when state is off
- Add batt_poll_failures tracking to alarm poll timeout (was missing,
  only analog poll timeout tracked failures)
- Add CAN tracking var resets to on_connect (limits, extremes, flags,
  decoded state, charge state, temps) for consistent re-publish behavior
- Remove dead code from set_include.h: Rs485BusyGuard, unused includes
- Fix can_frame_preamble type mismatch: unsigned long -> uint32_t

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

+++ firmware/deye-site/esphome/includes/set_include.h
@@ -1,7 +1,5 @@
 #pragma once
-#include <set>
 #include <string>
-#include <sstream>
 #include <vector>
 #include <cstdint>
 #include <cstdarg>
@@ -63,7 +61,7 @@ inline std::string rs485_make_cmd(int addr, int cid2, int batt_num) {
 
 // CAN frame processing helper - common preamble for all CAN handlers
 // Returns true if frame is valid (expected_size), false if invalid
-inline bool can_frame_preamble(const std::vector<uint8_t>& x, int& frame_count, unsigned long& last_rx, bool& stale, int& error_count, size_t expected_size = 8) {
+inline bool can_frame_preamble(const std::vector<uint8_t>& x, int& frame_count, uint32_t& last_rx, bool& stale, int& error_count, size_t expected_size = 8) {
     frame_count++;
     last_rx = millis();
     if (stale) { stale = false; }
